refactor: split markdown.ts, extract scroll manager, add polymorphic extractor settings
- Split 560-line markdown.ts into 4 modules with barrel re-export
  (markdown-rules, markdown-deep-research, markdown-formatting)
- Extract auto-scroll logic from gemini.ts to scroll-manager.ts
- Add polymorphic applySettings() to IConversationExtractor, replacing
  instanceof checks in applyExtractorSettings()
- Fix parameter naming in chatgpt.ts (_match convention)
- Add ADR-003, DES-012, WF-012 documentation

Zero consumer or test import changes. 742 tests pass.
