+++ src/content/markdown.ts
@@ -1,393 +1,26 @@
 /**
- * HTML to Markdown conversion using Turndown
+ * HTML to Markdown conversion — barrel re-export + orchestrator
+ *
+ * Internal modules:
+ * - markdown-rules.ts      — Turndown engine (leaf, no internal deps)
+ * - markdown-deep-research.ts — Citation → footnote pipeline
+ * - markdown-formatting.ts — Message formatting templates
  */
 
-import TurndownService from 'turndown';
+import { formatMessage, formatToolContent } from './markdown-formatting';
+import { convertDeepResearchContent } from './markdown-deep-research';
+import { generateHash } from '../lib/hash';
+import { MAX_FILENAME_BASE_LENGTH, FILENAME_ID_SUFFIX_LENGTH } from '../lib/constants';
 import type {
   ConversationData,
   ObsidianNote,
   NoteFrontmatter,
   TemplateOptions,
-  DeepResearchLinks,
-  DeepResearchSource,
 } from '../lib/types';
-import { generateHash } from '../lib/hash';
-import { buildSourceMap } from '../lib/source-map';
-import {
-  MAX_FILENAME_BASE_LENGTH,
-  FILENAME_ID_SUFFIX_LENGTH,
-  PLATFORM_LABELS,
-} from '../lib/constants';
 
-// ============================================================
-// Deep Research Link Processing Functions (Obsidian Footnote Mode)
-// ============================================================
-
-/** Pre-compiled regex for source-footnote wrapped citations */
-const CITATION_PATTERN_WRAPPED =
-  /<source-footnote[^>]*>[\s\S]*?<sup[^>]*?data-turn-source-index="(\d+)"[^>]*?>[\s\S]*?<\/sup>[\s\S]*?<\/source-footnote>/gi;
-
-/** Pre-compiled regex for standalone sup citations (fallback) */
-const CITATION_PATTERN_STANDALONE =
-  /<sup[^>]*?data-turn-source-index="(\d+)"[^>]*?>[\s\S]*?<\/sup>/gi;
-
-/**
- * Escape Markdown link metacharacters in text
- * Prevents injection of Markdown links via crafted titles
- */
-function escapeMarkdownLink(text: string): string {
-  return text.replace(/[\\[\]()]/g, '\\$&');
-}
-
-/**
- * Escape angle brackets in a single line of Markdown text.
- * Preserves blockquote markers and inline code segments.
- */
-function escapeAngleBracketsInLine(line: string): string {
-  // 1. Extract blockquote prefix (preserve as-is)
-  const bqMatch = line.match(/^(\s*>\s*)+/);
-  const prefix = bqMatch ? bqMatch[0] : '';
-  const rest = line.slice(prefix.length);
-
-  // 2. Split by inline code segments (capture group → odd indices are code)
-  const parts = rest.split(/(`[^`]+`)/);
-
-  // 3. Escape angle brackets in non-code segments.
-  //    Also handle \< and \> (backslash+angle) to prevent incomplete escaping (CodeQL js/incomplete-sanitization).
-  const escaped = parts
-    .map((part, i) => {
-      if (i % 2 === 1) return part; // inline code — preserve
-      return part.replace(/\\[<>]|[<>]/g, match => {
-        if (match.length === 2) {
-          // \< or \> → escaped backslash + escaped angle bracket
-          return '\\\\' + '\\' + match[1];
-        }
-        return '\\' + match;
-      });
-    })
-    .join('');
-
-  return prefix + escaped;
-}
-
-/**
- * Escape angle brackets in Markdown text for safe Obsidian rendering.
- * Preserves brackets inside fenced code blocks, inline code, and
- * blockquote markers.
- *
- * CommonMark §2.4: \< and \> are valid backslash escapes.
- */
-export function escapeAngleBrackets(text: string): string {
-  const lines = text.split('\n');
-  let inFencedBlock = false;
-
-  const result = lines.map(line => {
-    // Detect fenced code delimiter (may be inside a blockquote)
-    const stripped = line.replace(/^(\s*>\s*)*/, '');
-    if (/^`{3,}/.test(stripped)) {
-      inFencedBlock = !inFencedBlock;
-      return line;
-    }
-
-    if (inFencedBlock) return line;
-
-    return escapeAngleBracketsInLine(line);
-  });
-
-  return result.join('\n');
-}
-
-/**
- * Sanitize URL to remove dangerous schemes
- */
-function sanitizeUrl(url: string): string {
-  const dangerousSchemes = ['javascript:', 'data:', 'vbscript:', 'blob:'];
-  const lowerUrl = url.toLowerCase().trim();
-
-  for (const scheme of dangerousSchemes) {
-    if (lowerUrl.startsWith(scheme)) {
-      return ''; // Return empty for dangerous URLs
-    }
-  }
-
-  return url;
-}
-
-/**
- * Create a citation replacer callback for use with String.replace()
- * Shared logic for both wrapped and standalone citation patterns.
- */
-function createCitationReplacer(
-  sourceMap: Map<number, DeepResearchSource>
-): (_match: string, indexStr: string) => string {
-  return (_match: string, indexStr: string): string => {
-    const index = parseInt(indexStr, 10);
-    const source = sourceMap.get(index);
-    if (source) {
-      // Return placeholder span with content (Turndown filters empty elements)
-      // The custom Turndown rule will convert this to [^N]
-      return `<span data-footnote-ref="${index}">REF</span>`;
-    }
-    // Source not found: log warning and remove marker silently
-    console.warn(`[G2O] Citation reference ${index} not found in source map`);
-    return '';
-  };
-}
-
-/**
- * Convert inline citations to footnote reference placeholders
- *
- * Before: <source-footnote><sup data-turn-source-index="N">...</sup></source-footnote>
- * After: <span data-footnote-ref="N"></span>
- *
- * Design: We insert placeholder spans that survive Turndown processing,
- * then replace them with Obsidian footnote syntax [^N] after conversion.
- * This avoids double-escaping issues where Markdown gets re-escaped by Turndown.
- *
- * Important: data-turn-source-index is 1-based and may be non-sequential
- *
- * @param html HTML content to convert
- * @param sourceMap Map built from buildSourceMap()
- */
-function convertInlineCitationsToFootnoteRefs(
-  html: string,
-  sourceMap: Map<number, DeepResearchSource>
-): string {
-  const replacer = createCitationReplacer(sourceMap);
-
-  // Pattern 1: source-footnote wrapped
-  CITATION_PATTERN_WRAPPED.lastIndex = 0;
-  let result = html.replace(CITATION_PATTERN_WRAPPED, replacer);
-
-  // Pattern 2: standalone sup element (fallback)
-  CITATION_PATTERN_STANDALONE.lastIndex = 0;
-  result = result.replace(CITATION_PATTERN_STANDALONE, replacer);
-
-  return result;
-}
-
-/**
- * Generate References section with Obsidian footnote definitions
- *
- * Output format:
- * # References
- *
- * [^1]: [Title1](URL1)
- * [^2]: [Title2](URL2)
- * ...
- *
- * @param sources All sources from the source list (includes unreferenced sources)
- * @returns Markdown string for References section
- */
-function generateReferencesSection(sources: DeepResearchSource[]): string {
-  if (sources.length === 0) {
-    return '';
-  }
-
-  const lines: string[] = ['', '# References', ''];
-
-  sources.forEach((source, arrayIndex) => {
-    // data-turn-source-index is 1-based
-    const footnoteIndex = arrayIndex + 1;
-    const safeUrl = sanitizeUrl(source.url);
-
-    if (safeUrl) {
-      // [^N]: [Title](URL)
-      lines.push(`[^${footnoteIndex}]: [${escapeMarkdownLink(source.title)}](${safeUrl})`);
-    } else {
-      // URL invalid: title only
-      lines.push(`[^${footnoteIndex}]: ${escapeMarkdownLink(source.title)}`);
-    }
-  });
-
-  return lines.join('\n');
-}
-
-/**
- * Remove sources-carousel-inline elements
- */
-function removeSourcesCarousel(html: string): string {
-  return html.replace(/<sources-carousel-inline[\s\S]*?<\/sources-carousel-inline>/gi, '');
-}
-
-/**
- * Convert Deep Research content with Obsidian footnotes
- *
- * Design: Converts inline citations to [^N] footnote references
- * and generates a References section with footnote definitions.
- *
- * Processing flow:
- * 1. Build source map from links
- * 2. Convert <sup data-turn-source-index> to placeholder spans
- * 3. Remove sources carousel
- * 4. Convert HTML to Markdown (Turndown)
- * 5. Replace placeholder spans with [^N] footnote refs
- * 6. Append References section with footnote definitions
- */
-export function convertDeepResearchContent(html: string, links?: DeepResearchLinks): string {
-  let processed = html;
-
-  // 1. Build source map (1-based index)
-  let sourceMap = new Map<number, DeepResearchSource>();
-  if (links && links.sources.length > 0) {
-    sourceMap = buildSourceMap(links.sources);
-  }
-
-  // 2. Convert inline citations to placeholder spans
-  processed = convertInlineCitationsToFootnoteRefs(processed, sourceMap);
-
-  // 3. Remove sources carousel
-  processed = removeSourcesCarousel(processed);
-
-  // 4. Convert HTML to Markdown (Turndown rule converts spans to [^N])
-  const markdown = htmlToMarkdown(processed);
-
-  // 5. Add References section with all sources
-  if (links && links.sources.length > 0) {
-    return markdown + generateReferencesSection(links.sources);
-  }
-
-  return markdown;
-}
-
-// ============================================================
-// Turndown Configuration
-// ============================================================
-
-// Initialize Turndown with custom rules
-const turndown = new TurndownService({
-  headingStyle: 'atx',
-  codeBlockStyle: 'fenced',
-  bulletListMarker: '-',
-  emDelimiter: '*',
-  strongDelimiter: '**',
-});
-
-// Custom rule for code blocks with language detection
-turndown.addRule('codeBlocks', {
-  filter: node => {
-    return node.nodeName === 'PRE' && node.querySelector('code') !== null;
-  },
-  replacement: (content, node) => {
-    const codeElement = (node as HTMLElement).querySelector('code');
-    if (!codeElement) return content;
-
-    // Try to detect language from class
-    const className = codeElement.className || '';
-    const langMatch = className.match(/language-(\w+)/);
-    const lang = langMatch ? langMatch[1] : '';
-
-    const code = codeElement.textContent || '';
-    return `\n\`\`\`${lang}\n${code.trim()}\n\`\`\`\n`;
-  },
-});
-
-// Custom rule for inline code
-turndown.addRule('inlineCode', {
-  filter: node => {
-    return node.nodeName === 'CODE' && node.parentNode?.nodeName !== 'PRE';
-  },
-  replacement: content => {
-    return `\`${content}\``;
-  },
-});
-
-// Custom rule for footnote reference placeholders
-// Converts <span data-footnote-ref="N">REF</span> to [^N]
-turndown.addRule('footnoteRef', {
-  filter: node => {
-    return node.nodeName === 'SPAN' && (node as HTMLElement).hasAttribute('data-footnote-ref');
-  },
-  replacement: (_content, node) => {
-    const index = (node as HTMLElement).getAttribute('data-footnote-ref');
-    return `[^${index}]`;
-  },
-});
-
-// Custom rule for display math blocks (Gemini KaTeX: <div data-math="...">)
-turndown.addRule('mathBlock', {
-  filter: node => {
-    return node.nodeName === 'DIV' && (node as HTMLElement).hasAttribute('data-math');
-  },
-  replacement: (_content, node) => {
-    const latex = (node as HTMLElement).getAttribute('data-math');
-    if (!latex) return _content;
-    return `\n$$\n${latex}\n$$\n`;
-  },
-});
-
-// Custom rule for inline math (Gemini KaTeX: <span data-math="...">)
-turndown.addRule('mathInline', {
-  filter: node => {
-    return (
-      node.nodeName === 'SPAN' &&
-      (node as HTMLElement).hasAttribute('data-math') &&
-      !(node as HTMLElement).hasAttribute('data-footnote-ref')
-    );
-  },
-  replacement: (_content, node) => {
-    const latex = (node as HTMLElement).getAttribute('data-math');
-    if (!latex) return _content;
-    return `$${latex}$`;
-  },
-});
-
-// Custom rule for tables
-turndown.addRule('tables', {
-  filter: 'table',
-  replacement: (content, node) => {
-    const table = node as HTMLTableElement;
-    const rows: string[][] = [];
-
-    // Extract headers
-    const headerRow = table.querySelector('thead tr, tr:first-child');
-    if (headerRow) {
-      const headers: string[] = [];
-      headerRow.querySelectorAll('th, td').forEach(cell => {
-        headers.push(cell.textContent?.trim() || '');
-      });
-      if (headers.length > 0) {
-        rows.push(headers);
-      }
-    }
-
-    // Extract body rows
-    const bodyRows = table.querySelectorAll('tbody tr, tr:not(:first-child)');
-    bodyRows.forEach(row => {
-      const cells: string[] = [];
-      row.querySelectorAll('td, th').forEach(cell => {
-        cells.push(cell.textContent?.trim() || '');
-      });
-      if (cells.length > 0) {
-        rows.push(cells);
-      }
-    });
-
-    if (rows.length === 0) return '';
-
-    // Build markdown table
-    const lines: string[] = [];
-    rows.forEach((row, index) => {
-      lines.push('| ' + row.join(' | ') + ' |');
-      if (index === 0) {
-        // Add separator after header
-        lines.push('| ' + row.map(() => '---').join(' | ') + ' |');
-      }
-    });
-
-    return '\n' + lines.join('\n') + '\n';
-  },
-});
-
-/**
- * Convert HTML content to Markdown
- */
-export function htmlToMarkdown(html: string): string {
-  // Clean up HTML before conversion
-  const cleaned = html.replace(/<br\s*\/?>/gi, '\n').replace(/&nbsp;/g, ' ');
-
-  return escapeAngleBrackets(turndown.turndown(cleaned));
-}
+// Re-exports (preserve existing import paths)
+export { htmlToMarkdown, escapeAngleBrackets } from './markdown-rules';
+export { convertDeepResearchContent } from './markdown-deep-research';
 
 /**
  * Generate sanitized filename from title
@@ -408,95 +41,6 @@ export function generateFileName(title: string, conversationId: string): string 
  */
 export function generateContentHash(content: string): string {
   return generateHash(content);
-}
-
-/**
- * Get display label for AI assistant based on source platform
- */
-function getAssistantLabel(source: string): string {
-  return PLATFORM_LABELS[source] ?? 'Assistant';
-}
-
-/**
- * Format a single message according to template options
- */
-function formatMessage(
-  content: string,
-  role: 'user' | 'assistant',
-  options: TemplateOptions,
-  source: string
-): string {
-  // Convert HTML to Markdown for assistant messages; escape angle brackets for user messages
-  const markdown = role === 'assistant' ? htmlToMarkdown(content) : escapeAngleBrackets(content);
-  const assistantLabel = getAssistantLabel(source);
-
-  switch (options.messageFormat) {
-    case 'callout': {
-      const calloutType = role === 'user' ? options.userCalloutType : options.assistantCalloutType;
-      const label = role === 'user' ? 'User' : assistantLabel;
-      // Format as Obsidian callout with proper line handling
-      const lines = markdown.split('\n');
-      const formattedLines = lines.map((line, i) =>
-        i === 0 ? `> [!${calloutType}] ${label}\n> ${line}` : `> ${line}`
-      );
-      return formattedLines.join('\n');
-    }
-
-    case 'blockquote': {
-      const label = role === 'user' ? '**User:**' : `**${assistantLabel}:**`;
-      const lines = markdown.split('\n').map(line => `> ${line}`);
-      return `${label}\n${lines.join('\n')}`;
-    }
-
-    case 'plain':
-    default: {
-      const label = role === 'user' ? '**User:**' : `**${assistantLabel}:**`;
-      return `${label}\n\n${markdown}`;
-    }
-  }
-}
-
-/**
- * Format tool-use content as a collapsible callout or equivalent format
- *
- * Renders tool content (web search, code interpreter) as a separate block
- * before the assistant response message.
- *
- * @param toolContent Raw tool content string (may contain bold summary, queries, results)
- * @param options Template options for format selection
- */
-function formatToolContent(toolContent: string, options: TemplateOptions): string {
-  const lines = toolContent.split('\n').filter(l => l.trim());
-
-  // Extract first bold line as callout title (e.g., "**Searched the web**" → "Searched the web")
-  let title = 'Tool Activity';
-  let bodyLines = lines;
-  if (lines[0]?.startsWith('**') && lines[0]?.endsWith('**')) {
-    title = lines[0].slice(2, -2);
-    bodyLines = lines.slice(1);
-  }
-
-  switch (options.messageFormat) {
-    case 'callout': {
-      // Collapsible callout: [!ABSTRACT]- collapsed by default
-      if (bodyLines.length === 0) {
-        return `> [!ABSTRACT]- ${title}`;
-      }
-      const formatted = bodyLines.map(line => `> ${line}`);
-      return `> [!ABSTRACT]- ${title}\n${formatted.join('\n')}`;
-    }
-
-    case 'blockquote': {
-      const header = `**${title}**`;
-      const quoted = bodyLines.map(line => `> ${line}`);
-      return quoted.length > 0 ? `${header}\n${quoted.join('\n')}` : header;
-    }
-
-    case 'plain':
-    default: {
-      return bodyLines.length > 0 ? `**${title}**\n${bodyLines.join('\n')}` : `**${title}**`;
-    }
-  }
 }
 
 /**
